Done and tested

+++ src/main/java/ru/yandex/practicum/filmorate/storage/UserDbStorage.java
@@ -1,6 +1,5 @@
 package ru.yandex.practicum.filmorate.storage;
 
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,12 +10,13 @@ import ru.yandex.practicum.filmorate.model.*;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.ArrayList;
 import java.util.List;
 
 @Component
 public class UserDbStorage implements UserStorage {
 
-    private final static Logger log = LoggerFactory.getLogger(InMemoryFilmStorage.class);
+    private final static Logger log = LoggerFactory.getLogger(UserDbStorage.class);
     private final JdbcTemplate jdbcTemplate;
 
     @Autowired
@@ -33,14 +33,14 @@ public class UserDbStorage implements UserStorage {
 
     @Override
     public boolean checkFriendshipExist(Integer userId, Integer friendId) {
-        String sqlQuery = "select count(*) from friendship_approved where user_id = ?, friend_id = ?";
+        String sqlQuery = "select count(*) from friendship_approved where user_id = ? and friend_id = ?";
         Integer result = jdbcTemplate.queryForObject(sqlQuery, Integer.class, userId, friendId);
         return result == 1;
     }
 
     @Override
     public boolean checkFriendshipRequestedByUser(Integer userId, Integer friendId) {
-        String sqlQuery = "select count(*) from friendship_requests where user_id = ?, friend_id = ?";
+        String sqlQuery = "select count(*) from friendship_requests where user_id = ? and friend_id = ?";
         Integer result = jdbcTemplate.queryForObject(sqlQuery, Integer.class, userId, friendId);
         return result == 1;
     }
@@ -64,7 +64,7 @@ public class UserDbStorage implements UserStorage {
 
     @Override
     public User getUser(int id) {
-        String sqlQuery = "select user_id, name, email, login, birthdate" +
+        String sqlQuery = "select user_id, name, email, login, birthdate " +
                 "from users where user_id = ?";
 
         return jdbcTemplate.queryForObject(sqlQuery, this::mapRowToUser, id);
@@ -72,7 +72,7 @@ public class UserDbStorage implements UserStorage {
 
     @Override
     public User updateUser(User user) {
-        String sqlQuery = "update users set " + "name = ?, email = ?, login = ?, birthdate = ?" + "where id = ?";
+        String sqlQuery = "update users set " + "name = ?, email = ?, login = ?, birthdate = ?" + " where user_id = ?";
         jdbcTemplate.update(sqlQuery
                 , user.getName()
                 , user.getEmail()
@@ -91,41 +91,61 @@ public class UserDbStorage implements UserStorage {
     }
 
     @Override
-    public void addFriend(Integer userId, Integer friendId) {
-        if (checkFriendshipRequestedByUser(friendId, userId)) {
+    public List<User> addFriend(Integer userId, Integer friendId) {
+        if (checkFriendshipRequestedByUser(userId, friendId)) {
 
-            removeFriendshipRequest(friendId, userId);
-            approveFriendShip(friendId, userId);
+            removeFriendshipRequest(userId, friendId);
+            approveFriendShip(userId, friendId);
 
             log.info("Friendship approved");
+
+            return getAllFriendsOfUser(userId);
         }
 
-        String sqlQuery = "insert into friendship_requests (user_id, friend_id)" +
+        String sqlQuery = "insert into friendship_requests (user_id, friend_id) " +
                 "values (?, ?)";
-        jdbcTemplate.update(sqlQuery, userId, friendId);
+        jdbcTemplate.update(sqlQuery, friendId, userId);
 
-        approveFriendShip(friendId, userId);
+        approveFriendShip(userId, friendId);
 
-        log.info("Friendship requested");
+        log.info("Friendship requested by friend");
+        return getAllFriendsOfUser(userId);
     }
 
     @Override
     public List<User> getAllFriendsOfUser(Integer userId) {
-        String sqlQuery = "select friendship_approved.friend_id, users.name, users.email, users.login, users.birthdate" +
-                "from friendship_approved" + "join users on friendship_approved.friend_id = users.user_id" +
-                "where friendship_approved.user_id = ?";
+        List<Integer> friendsId = getAllFriendsId(userId);
+        List<User> friends = new ArrayList<>();
+        for (Integer id : friendsId) {
+            friends.add(getUser(id));
+        }
+        return friends;
+    }
 
-        return jdbcTemplate.query(sqlQuery, this::mapRowToUser, userId);
+
+    @Override
+    public List<Integer> getAllFriendsId(Integer userId) {
+        /*String sqlQuery = "select f.friend_id, u.name, u.email, u.login, u.birthdate " +
+                "from friendship_approved as f " + "join users as u on f.friend_id = u.user_id " +
+                "where f.user_id = ?";
+
+        return jdbcTemplate.query(sqlQuery, this::mapRowToUser, userId);*/
+
+        String sqlQuery = "select friend_id " + "from friendship_approved " +
+                "where user_id = ?";
+
+        return jdbcTemplate.queryForList(sqlQuery, Integer.class, userId);
     }
 
     @Override
     public void removeUserFromFriends(Integer userId, Integer friendId) {
 
-        if (!checkIfUserInFriendsList(friendId, userId)) {
+        if (!checkIfUserInFriendsList(userId, friendId)) {
             log.info("There`s no such friend of user");
         }
 
-        if (checkFriendshipRequestedByUser(userId, friendId)) {
+        if (checkFriendshipRequestedByUser(friendId, userId)) {
+
             String sqlQuery = "delete from friendship_approved where user_id = ? and friend_id = ?";
             jdbcTemplate.update(sqlQuery, userId, friendId);
             log.info("Friend removed");
@@ -135,7 +155,7 @@ public class UserDbStorage implements UserStorage {
             log.info("Friendship Request declined");
         }
 
-        if (checkFriendshipExist(userId, friendId)) {
+        if (checkFriendshipExist(userId, friendId) && checkFriendshipExist(friendId, userId)) {
             String sqlQuery = "delete from friendship_approved where user_id = ? and friend_id = ?";
             jdbcTemplate.update(sqlQuery, userId, friendId);
             log.info("Friend removed");
@@ -144,9 +164,6 @@ public class UserDbStorage implements UserStorage {
             jdbcTemplate.update(sqlQuery, friendId, userId);
             log.info("Friendship removed");
         }
-
-        String sqlQuery = "insert into friendship_requests (user_id, friend_id)" + "values (?, ?)";
-        jdbcTemplate.update(sqlQuery, userId, friendId);
     }
 
     @Override
@@ -167,6 +184,15 @@ public class UserDbStorage implements UserStorage {
                 "values (?, ?)";
         jdbcTemplate.update(sqlQuery, userId, friendId);
     }
+
+    @Override
+    public void addFriendshipRequest(Integer userId, Integer friendId) {
+
+        String sqlQuery = "insert into friendship_requests (user_id, friend_id)" +
+                "values (?, ?)";
+        jdbcTemplate.update(sqlQuery, userId, friendId);
+    }
+
 
     @Override
     public User mapRowToUser(ResultSet resultSet, int rowNum) throws SQLException {
